# backend/src/main/resources/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MeerKatters API
  description: |
    API REST para la plataforma MeerKatters - Comunidad de estudiantes para 
    organización de actividades de estudio colaborativo.
    
    ## Modelo de Autorización
    
    **IMPORTANTE**: Un mismo usuario puede ser tanto estudiante como tutor de distintos cursos/comunidades.
    Los permisos se evalúan por contexto específico del recurso (usuarioId + comunidadId/eventoId),
    NO dependen exclusivamente de un rol global estático.
    
    ### Roles por Contexto:
    - **Admin de Comunidad**: Usuario que creó la comunidad o recibió transferencia de admin
    - **Miembro**: Usuario que pertenece a una comunidad
    - **Tutor contratado**: Tutor asociado a un grupo específico
    - **Creador de Evento**: Usuario que creó un evento específico
    
    ## Autenticación
    
    La API utiliza **JWT (JSON Web Tokens)** para autenticación.
    Incluir el token en el header `Authorization: Bearer <token>`
    
  version: 1.0.0
  contact:
    name: Equipo MeerKatters
    email: meerkatters@example.com
  license:
    name: MIT
servers:
  - url: https://api.meerkatters.com/v1
    description: Servidor de producción
  - url: https://staging-api.meerkatters.com/v1
    description: Servidor de staging
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local
...
paths:
...
  /users/me:
    get:
      tags:
        - Usuarios
      summary: Obtener perfil propio
      description: Devuelve el perfil del usuario autenticado
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Usuarios
      summary: Actualizar perfil propio
      description: Actualiza los datos del perfil del usuario autenticado
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Usuarios
      summary: Eliminar cuenta propia
      description: Elimina la cuenta del usuario autenticado. Esta acción es irreversible.
      operationId: deleteCurrentUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cuenta eliminada exitosamente
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/password:
    put:
      tags:
        - Usuarios
      summary: Cambiar contraseña
      description: Cambia la contraseña del usuario autenticado
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Contraseña actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/visibility:
    put:
      tags:
        - Usuarios
      summary: Configurar visibilidad en listados
      description: Configura si el perfil aparece en listados públicos
      operationId: updateVisibility
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisibilityRequest'
      responses:
        '200':
          description: Visibilidad actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{userId}:
    get:
      tags:
        - Usuarios
      summary: Obtener perfil de usuario
      description: Devuelve el perfil público de un usuario específico
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /auth/login
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: ID del usuario
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TipoPlan:
      type: string
      enum:
        - FREE
        - PREMIUM
      description: Tipo de plan de usuario
...
    UserDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        nombre:
          type: string
        foto:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        intereses:
          type: array
          items:
            type: string
        googleId:
          type: string
          nullable: true
        visibleEnListados:
          type: boolean
        plan:
          $ref: '#/components/schemas/TipoPlan'
        esTutor:
          type: boolean
          description: Indica si el usuario tiene perfil de Tutor
        createdAt:
          type: string
          format: date-time
    UserPublicResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        foto:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        intereses:
          type: array
          items:
            type: string
        esTutor:
          type: boolean
    UpdateUserRequest:
      type: object
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        foto:
          type: string
          format: uri
        bio:
          type: string
          maxLength: 500
        intereses:
          type: array
          items:
            type: string
          maxItems: 10
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
    VisibilityRequest:
      type: object
      required:
        - visibleEnListados
      properties:
        visibleEnListados:
          type: boolean

    # =====================================================
    # Tutor DTOs
    # =====================================================
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Código de error
        message:
          type: string
          description: Mensaje descriptivo del error
        details:
          type: array
          items:
            type: string
          description: Detalles adicionales del error
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: Ruta del endpoint que generó el error
...